+++ src/mongo/db/repl/tenant_database_cloner.cpp
@@ -322,8 +322,8 @@ BSONObj TenantDatabaseCloner::Stats::toBSON() const {
 }
 
 void TenantDatabaseCloner::Stats::append(BSONObjBuilder* builder) const {
-    builder->appendNumber("collections", collections);
-    builder->appendNumber("clonedCollections", clonedCollections);
+    builder->appendNumber("collections", static_cast<long long>(collections));
+    builder->appendNumber("clonedCollections", static_cast<long long>(clonedCollections));
     if (start != Date_t()) {
         builder->appendDate("start", start);
         if (end != Date_t()) {
